Create readable-with component
This component aims to abstract the positioning of elements when not
too much space is available.  Its main goal is to have a clean way of
rendering contents somewhat more to the left side of the screen.

It's by no means perfect as it stands, but the abstraction may help us
clear up other things as we move along.